# openapi2.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Hobotti API",
    "description": "A basic backend API for the hobby/likemided community finder mobile app Hobotti. <br> Please contact the developers to gain access to the API key needed to use this API.",
    "contact": {
      "email": "santtu-taskinen@outlook.com"
    }
  },
...
  "paths": {
...
    "/users": {
      "post": {
        "security": [
          {
            "APIAccess": [],
            "HobottiAccess": []
          }
        ],
        "tags": [
          "Users"
        ],
        "summary": "Register an user",
        "description": "Create new user in system. Takes in the APIkey in the \"hobotti-api-key\" header.",
        "responses": {
          "200": {
            "description": "New user is created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/authsuccess"
                }
              }
            }
          },
          "403": {
            "description": "No valid email provided",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/messageResponse"
                }
              }
            }
          },
          "463": {
            "description": "No API Key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/messageResponse"
                }
              }
            }
          },
          "464": {
            "description": "Invalid API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/messageResponse"
                }
              }
            }
          },
          "473": {
            "description": "No access token provided.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/messageResponse"
                }
              }
            }
          },
          "500": {
            "description": "Something went wrong",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/messageResponse"
                }
              }
            }
          },
          "565": {
            "description": "Something went wrong with the API key verification.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/messageResponse"
                }
              }
            }
          },
          "570": {
            "description": "Something went wrong while trying to authenticate token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/messageResponse"
                }
              }
            }
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/users"
              }
            }
          },
          "description": "User that we want to create"
        }
      }
    },
    "/users/login": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "login to the system",
        "description": "Login into the system",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/login"
              }
            }
          },
          "description": "Login Credentials"
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/authsuccess"
                }
              }
            }
          },
          "401": {
            "description": "Wrong password was given",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/messageResponse"
                }
              }
            }
          },
          "404": {
            "description": "User not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/messageResponse"
                }
              }
            }
          },
          "463": {
            "description": "No API Key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/messageResponse"
                }
              }
            }
          },
          "464": {
            "description": "Invalid API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/messageResponse"
                }
              }
            }
          },
          "473": {
            "description": "No access token provided.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/messageResponse"
                }
              }
            }
          },
          "500": {
            "description": "Something went wrong.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/messageResponse"
                }
              }
            }
          },
          "565": {
            "description": "Something went wrong with the API key verification.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/messageResponse"
                }
              }
            }
          },
          "570": {
            "description": "Something went wrong while trying to authenticate token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/messageResponse"
                }
              }
            }
          }
        }
      }
    },
    "/users/me": {
      "get": {
        "security": [
          {
            "APIAccess": [],
            "HobottiAccess": []
          }
        ],
        "tags": [
          "Users"
        ],
        "summary": "Get user data",
        "responses": {
          "200": {
            "description": "User was found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/users"
                }
              }
            }
          },
          "404": {
            "description": "No user found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/messageResponse"
                }
              }
            }
          },
          "463": {
            "description": "No API Key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/messageResponse"
                }
              }
            }
          },
          "464": {
            "description": "Invalid API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/messageResponse"
                }
              }
            }
          },
          "473": {
            "description": "No access token provided.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/messageResponse"
                }
              }
            }
          },
          "500": {
            "description": "There was a problem finding the user.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/messageResponse"
                }
              }
            }
          },
          "565": {
            "description": "Something went wrong with the API key verification.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/messageResponse"
                }
              }
            }
          },
          "570": {
            "description": "Something went wrong while trying to authenticate token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/messageResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "security": [
          {
            "APIAccess": [],
            "HobottiAccess": []
          }
        ],
        "summary": "Delete user",
        "tags": [
          "Users"
        ],
        "responses": {
          "200": {
            "description": "User is deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/users"
                }
              }
            }
          },
          "463": {
            "description": "No API Key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/messageResponse"
                }
              }
            }
          },
          "464": {
            "description": "Invalid API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/messageResponse"
                }
              }
            }
          },
          "473": {
            "description": "No access token provided.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/messageResponse"
                }
              }
            }
          },
          "500": {
            "description": "Something went wrong with the user account deletion."
          },
          "565": {
            "description": "Something went wrong with the API key verification.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/messageResponse"
                }
              }
            }
          },
          "570": {
            "description": "Something went wrong while trying to authenticate token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/messageResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "security": [
          {
            "APIAccess": [],
            "HobottiAccess": []
          }
        ],
        "summary": "Update user",
        "tags": [
          "Users"
        ],
        "responses": {
          "200": {
            "description": "User is updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/users"
                }
              }
            }
          },
          "404": {
            "description": "User not found."
          },
          "463": {
            "description": "No API Key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/messageResponse"
                }
              }
            }
          },
          "464": {
            "description": "Invalid API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/messageResponse"
                }
              }
            }
          },
          "473": {
            "description": "No access token provided.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/messageResponse"
                }
              }
            }
          },
          "500": {
            "description": "Something went wrong while updating the user."
          },
          "565": {
            "description": "Something went wrong with the API key verification.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/messageResponse"
                }
              }
            }
          },
          "570": {
            "description": "Something went wrong while trying to authenticate token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/messageResponse"
                }
              }
            }
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/users"
              }
            }
          },
          "description": "User with new values of properties"
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "https://hobotti-backend-testing.herokuapp.com/api/"
    },
    {
      "url": "http://hobotti-backend-testing.herokuapp.com/api/"
    },
    {
      "url": "http://localhost:3000/api/"
    }
  ],
  "components": {
    "securitySchemes": {
      "APIAccess": {
        "type": "apiKey",
        "in": "header",
        "name": "hobotti-api-key"
      },
      "HobottiAccess": {
        "type": "apiKey",
        "in": "header",
        "name": "hobotti-access-token"
      }
    },
    "schemas": {
...
      "authsuccess": {
        "properties": {
          "token": {
            "type": "string"
          }
        }
      },
...
      "login": {
        "required": [
          "email",
          "password"
        ],
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "users": {
        "required": [
          "name",
          "email",
          "password",
          "hobbies"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "uniqueItems": true
          },
          "password": {
            "type": "string",
            "uniqueItems": true
          },
          "hobbies": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "messageResponse": {
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "APIAccess": [],
      "HobottiAccess": []
    }
  ]
}